requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
Django==4.0.5
django-cors-headers==3.13.0
django-phonenumber-field==6.3.0
djangorestframework==3.13.1
djlint==1.3.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.3
importlib-metadata==4.12.0
pathspec==0.9.0
phonenumbers==8.12.50
Pillow==9.1.1
PyJWT==2.4.0
PyMySQL==1.0.2
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
regex==2022.6.2
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tqdm==4.64.0
twilio==7.9.3
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
